Added more inline styling params aside from width
Due to width and fullwidth being the only two params that were affected
in the erronously named "sizes" list it was hard to see how we would
scale up style props to make a more personal button/anchor tag.

This change makes it so that the sizes list is now named style, and has
a height and a color that is hard coded to test functionality. The style
of the component now takes this list, and then overwrites the width if
fullWidth is enabled.